PIN added

+++ player.sh
@@ -16,11 +16,13 @@ function install
 	keystorePath=$1
 	keyID=$2
 	password=$3 
+	PIN=$4
+		
 		
 	echo "$1" >> conf
 	echo "$2" >> conf
 	echo "$3" >> conf
-	
+	echo "$4" >> conf
 	
 
 	#run keystore app to encrypt conf file
@@ -30,10 +32,18 @@ function install
 
 
 if [ -f "conf" ]; then
+
 	echo "Conf exist" 
 	
+	if [ $# -lt 1 ];then	
+		echo "Not enough arguments"
+		exit
+	fi
+	
 	filePath=$1
-	moment=$2
+	PIN=$2
+	echo "PIN "$PIN
+	moment=$3
 
 	"$keystorePath/exe" "decrypt" "/home/stas/ClionProjects/Kryptografia2.2/conf" $key
 	
@@ -41,12 +51,19 @@ if [ -f "conf" ]; then
 	kPath=`sed -n 1p conf`
 	kID=`sed -n 2p conf`
 	kPass=`sed -n 3p conf`
+	echo $kPass
+	kPIN=`sed -n 4p conf`
+	echo $kPIN
+	#"$keystorePath/exe" "encrypt" "/home/stas/ClionProjects/Kryptografia2.2/conf" $key
 	
-	"$keystorePath/exe" "encrypt" "/home/stas/ClionProjects/Kryptografia2.2/conf" $key
+	if [ "$kPIN" != "$PIN" ];then	
+		echo "Wrong PIN, bye, bye"
+		exit
+	fi
 	
 	echo "$kPass" >> pass
 		
-	#run keystore app with key ID
+	#run keystore app with key ID 	
 	"$keystorePath/exe" $kID < pass
 
 	rm pass
@@ -74,7 +91,7 @@ elif [ "$1" == install ]; then
 		exit
   	fi
   	
-  	install "$2" "$3" "$4"		#keystore_path, key_id, password
+  	install "$2" "$3" "$4" "$5"		#keystore_path, key_id, password
   	exit
 else
   echo "File not found"
@@ -89,33 +106,6 @@ if [ $# -lt 2 ];then
 	echo "Not enough arguments"
 	exit
 fi
-
-
-keystorePath=$1
-keyID=$2
-filePath=$3
-action=$4
-mode=$5
-
-
-
-
-#compile krypto2.1
-g++ main.cpp -lcrypto -lssl -o exe
-
-
-#run keystore app with key ID
-"$keystorePath/exe" $keyID
-
-
-#take key from temp file
-key=`cat "$keystorePath/temp"`
-rm "$keystorePath/temp"
-
-#echo $key
-
-#run krypto app with filePath, action, mode and key
-./exe $filePath $action $mode $key
